// src/app/shared/models/nutritional-data.ts
export class NutritionalData {
    // Macros
    energyKcal: number = 0;
    protein: number = 0;
    carbohydrate: number = 0;
    fat: number = 0;

    // Basic nutrients
    fiber: number = 0;
    sugar: number = 0;
    salt: number = 0;
    saturatedFat: number = 0;
    sodium: number = 0;
    cholesterol: number = 0;
    transFat: number = 0;
    omega3: number = 0;
    omega6: number = 0;
    omega9: number = 0;
    alcohol: number = 0;

    // Minerals
    calcium: number = 0;
    iron: number = 0;
    magnesium: number = 0;
    phosphorus: number = 0;
    potassium: number = 0;
    zinc: number = 0;
    copper: number = 0;
    manganese: number = 0;
    selenium: number = 0;
    iodine: number = 0;

    // Vitamins
    vitaminA: number = 0;
    vitaminD: number = 0;
    vitaminE: number = 0;
    vitaminK: number = 0;
    vitaminC: number = 0;
    vitaminB1: number = 0;
    vitaminB2: number = 0;
    vitaminB3: number = 0;
    vitaminB5: number = 0;
    vitaminB6: number = 0;
    vitaminB9: number = 0;
    vitaminB12: number = 0;
    biotin: number = 0;

    // Others
    caffeine: number = 0;
    taurine: number = 0;
}